HT3C_App/requirements.txt
appdirs==1.4.3
asgi-redis==1.4.3
asgiref==2.3.0
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
bbcode==1.0.31
bcrypt==3.1.4
boto==2.48.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
codecov==2.0.15
confusable-homoglyphs==3.0.0
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.1
cryptography==2.2.2
Django==2.0.4
django-admin-views==0.8.0
django-adminlte2==0.3.0
django-adminplus==0.5
django-annoying==0.10.4
django-appconf==1.0.2
django-bootstrap-form==3.4
django-braces==1.13.0
django-bulk-update==2.2.0
django-chartit==0.2.9
django-chat==0.1.0
django-contrib-comments==1.8.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-eventstream==2.0.1
django-filter==1.1.0
django-frontend-template==1.2.0
django-grip==1.8.0
django-guardian==1.4.9
django-jasmine==0.5.1
django-mathfilters==0.4.0
django-model-utils==3.1.1
django-oauth-toolkit==1.1.0
django-pagedown==1.0.4
django-private-chat==0.1.8
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-registration==2.4.1
django-registration-redux==2.4
django-selectable==1.1.0
django-simple-captcha==0.5.6
django-tables2==1.0.7
django-test-plus==1.0.22
django-threadedcomments==1.2
django-user-accounts==2.0.3
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djappsettings==0.4.0
drf-json-api==0.1.1
futures==3.1.1
geolocation-python==0.2.2
gripcontrol==3.2.1
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
msgpack-python==0.5.6
mysqlclient==1.3.12
nose==1.3.7
oauthlib==2.0.7
olefile==0.45.1
packaging==17.1
paramiko==2.4.1
pbr==4.0.2
Pillow==5.1.0
pinax-theme-bootstrap==8.0.1
pisa==3.0.33
pluggy==0.6.0
postmarkup==1.2.2
psycopg2==2.7.4
pubcontrol==2.4.2
py==1.5.3
pyasn1==0.4.2
pybbm==0.19.0
pycparser==2.18
PyJWT==1.6.1
PyNaCl==1.2.1
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
pytils==0.3
pytz==2018.4
RapidSMS==0.22.0
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
storage==0.0.4.3
tornado==5.0.2
tornado-psycopg2==0.3.1
tox==3.0.0
txaio==2.10.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
uuid==1.30
virtualenv==15.2.0
webencodings==0.5.1
websockets==3.2
Werkzeug==0.14.1
xhtml2pdf==0.2.2
xhtml2pdf-legacy==3.0.38
zope.interface==4.5.0
